Fix menu and post-login actions
